requirement.txt
decorator==4.0.11
Django==1.10.8
django-cors-headers==2.0.2
django-crispy-forms==1.6.0
django-filter==0.12.0
django-pgjsonb==0.0.16
django-ses==0.7.1
djangorestframework==3.6.4
djangorestframework-csv==1.4.1
djangorestframework-jwt==1.7.2
ipython==5.3.0
ipython-genutils==0.2.0
django-cors-headers==2.0.2
djangorestframework-jwt==1.7.2
